Extract configuration to constant
