+++ qgreenland/constants.py
@@ -1,6 +1,8 @@
 import os
 from enum import Enum
 
+import yamale
+
 from qgreenland import __version__
 
 PROJECT = 'qgreenland'
@@ -11,7 +13,6 @@ ENVIRONMENT = os.environ.get('ENVIRONMENT', 'dev')
 PACKAGE_DIR = os.path.dirname(os.path.realpath(__file__))
 DATA_DIR = '/luigi/data'
 WIP_DIR = f'{DATA_DIR}/luigi-wip'
-DATA_RELEASE_DIR = f'{DATA_DIR}/release/{__version__}'
 ASSETS_DIR = f'{PACKAGE_DIR}/assets'
 CONFIG_DIR = f'{PACKAGE_DIR}/config'
 CONFIG_SCHEMA_DIR = f'{CONFIG_DIR}/schema'
@@ -45,6 +46,37 @@ PROJECT_CRS = 'EPSG:3411'
 
 # URS stuff
 URS_COOKIE = 'urs_user_already_logged'
+
+
+def _load_config(config_filename):
+    """Validate config file against schema with Yamale.
+
+    It is expected that the given config filename in CONFIG_DIR has a schema of
+    matching name in CONFIG_SCHEMA_DIR.
+
+    Yamale can read in directories of config files, so it returns a list of
+    (data, fp) tuples. We always read single files, so we return just the data
+    from result[0][0].
+    """
+    config_fp = os.path.join(CONFIG_DIR, config_filename)
+    schema_fp = os.path.join(CONFIG_SCHEMA_DIR, config_filename)
+
+    if not os.path.isfile(config_fp):
+        return NotImplementedError(
+            'Loading is supported for only one config file at a time.'
+        )
+
+    schema = yamale.make_schema(schema_fp)
+    config = yamale.make_data(config_fp)
+    yamale.validate(schema, config)
+
+    return config[0][0]
+
+
+CONFIG = {
+    'layers': _load_config('layers.yml'),
+    'layer_groups': _load_config('layer_groups.yml')
+}
 
 
 class TaskType(Enum):
