+++ qgreenland/util/misc.py
@@ -3,10 +3,7 @@ import shutil
 import time
 from contextlib import contextmanager
 
-import yamale
-
-from qgreenland.constants import (CONFIG_DIR,
-                                  CONFIG_SCHEMA_DIR,
+from qgreenland.constants import (CONFIG,
                                   DATA_RELEASE_DIR,
                                   REQUEST_TIMEOUT,
                                   TaskType,
@@ -83,50 +80,18 @@ def cleanup_output_dirs(delete_fetch_dir=False):
             _rmtree(directory)
 
 
-def _load_config(config_filename):
-    """Validate config file against schema with Yamale.
-
-    It is expected that the given config filename in CONFIG_DIR has a schema of
-    matching name in CONFIG_SCHEMA_DIR.
-
-    Yamale can read in directories of config files, so it returns a list of
-    (data, fp) tuples. We always read single files, so we return just the data
-    from result[0][0].
-    """
-    config_fp = os.path.join(CONFIG_DIR, config_filename)
-    schema_fp = os.path.join(CONFIG_SCHEMA_DIR, config_filename)
-
-    if not os.path.isfile(config_fp):
-        return NotImplementedError(
-            'Loading is supported for only one config file at a time.'
-        )
-
-    schema = yamale.make_schema(schema_fp)
-    config = yamale.make_data(config_fp)
-    yamale.validate(schema, config)
-
-    return config[0][0]
-
-
-def load_layer_config(layername=None):
-    config = _load_config('layers.yml')
+def get_layer_config(layername=None):
+    config = CONFIG['layers']
 
     if not layername:
         return config
 
-    # TODO: Add error handling
     try:
         return config[layername]
     except KeyError:
         raise NotImplementedError(
             f"Configuration for layer '{layername}' not found."
         )
-
-
-def load_group_config():
-    config = _load_config('layer_groups.yml')
-
-    return config
 
 
 def get_layer_fs_path(layer_name, layer_cfg):
